Update queue_family_index to being uint32_t

+++ vulkan-cpp/types.hpp
@@ -476,7 +476,7 @@ namespace vk {
         std::span<const char*>
           extensions{}; // Can add VK_KHR_SWAPCHAIN_EXTENSION_NAME to this
                         // extension
-        uint8_t queue_family_index = 0;
+        uint32_t queue_family_index = 0;
     };
 
     // raw image handlers
